Penambahan Validasi Buka Tutup API Bulanan dari status pada tabel pendahuluan bulanan

+++ application/models/api_bulanan_model/Arus_kas_model.php
@@ -56,14 +56,22 @@ class Arus_kas_model extends CI_Model {
       $arrID[] = $value['id_aruskas'];
     }
 
+    $status_input = get_status_input($user,$tahun,$bulan);
+
+    if($status_input == true){
       
       $this->db->where('id_bulan',$bulan);
       $this->db->where('tahun',$tahun);
       $this->db->where('iduser',$user);
       $this->db->where_in('id_aruskas',$arrID);
       $this->db->delete($this->table);
-      return $this->db->affected_rows();
-    
+      $res['msg'].= $this->db->affected_rows()." data deleted, ";
+
+    } else {
+      $res['msg'].= "Invalid status input $user $tahun $bulan, ";
+
+    }
+    return $res;
   }
 
   public function insert($data,$tkn)
@@ -75,7 +83,7 @@ class Arus_kas_model extends CI_Model {
 
     $this->db->select('id_aruskas');
     $id = $this->db->get_where('mst_aruskas',array('mst_aruskas.iduser'=>$tknid))->result_array();
-    // $id = $this->db->get('mst_aruskas')->result_array();
+
     $arrID = array();
     foreach ($id as $key => $value) {
       $arrID[] = $value['id_aruskas'];
@@ -94,41 +102,46 @@ class Arus_kas_model extends CI_Model {
       $id_bulan = $value['id_bulan'];
       $id_user = $value['iduser'];
       $tahun = $value['tahun'];
-      
 
       // cek data id investasi
       if (in_array($id_aruskas, $arrID) && in_array($id_user, $arrUSER) && in_array($id_bulan, $arrBulan)) {
         // jika key nya not null maka id investasi merupakan hasil investasi
-        
-        $cekdata = $this->db->get_where($this->table,array('iduser'=>$id_user,'id_aruskas'=>$id_aruskas,'id_bulan'=>$id_bulan,'tahun'=>$tahun))->num_rows();
-        
-        if ($cekdata>0) {
 
-          // update
-          $dataUpdate[]=$value;
-        }else{
-          $dataInsert[]=$value;
-          // insert
-          
+        $status_input = get_status_input($id_user,$tahun,$id_bulan);
+
+        if($status_input == true){
+
+          $cekdata = $this->db->get_where($this->table,array('iduser'=>$id_user,'id_aruskas'=>$id_aruskas,'id_bulan'=>$id_bulan,'tahun'=>$tahun))->num_rows();
+        
+          if ($cekdata>0) {
+
+            $dataUpdate[]=$value;
+
+          }else{
+
+            $dataInsert[]=$value;
+            
+          }
+
+        } else {
+          $status = 0;
+          $res['msg'].="Invalid status input $id_user $tahun $id_bulan, ";
         }
 
       }else{
         $status = 0;
+        $res['msg'].="Invalid Id Aruskas $id_aruskas, ";
         // jika key nya null maka error karna bukan hasil investasi
       }
     }
 
-    $msg='| ';
-    $res=array();
-    if ($status==0) {
-      $res['error']=true;
-      $res['msg']='Data Invalid';
-    }else{
+    if ($status==1) 
+    {
       if ((is_countable($dataInsert)?$dataInsert:[])) { 
         // jika ada data yg diinput
         $this->db->insert_batch($this->table, $dataInsert);
         $jumlahInsert = $this->db->affected_rows();
-        $msg.= $jumlahInsert.' Data Berhasil Ditambahkan | ';
+        $msg.= $jumlahInsert.' data added, ';
       }
 
       if ((is_countable($dataUpdate)?$dataUpdate:[])) { 
@@ -144,19 +157,17 @@ class Arus_kas_model extends CI_Model {
           $this->db->update($this->table , $dataUpdateRow);
           $jumlahUpdate = $this->db->affected_rows();
           $jumlahUpdateAll = $jumlahUpdateAll+$jumlahUpdate;
-        }
-        // jika ada data yg diinput
-        
-        $msg.= $jumlahUpdateAll.' Data Berhasil Diperbarui | ';
-      }
 
+        }
+        $msg.= $jumlahUpdateAll.' data updated, ';
+      }
+      $res=array();
       $res['error']=false;
-      $res['msg']=$msg;
+      $res['msg'].=$msg;
     }
-    
-    
 
     return $res;
+
   }
 
 }
